Change the collect quest in order to be 100% atomic
Then it's no longer necessary to wait on collect because the collect
call is 100% sync. No more yield / command then the js event loop can
continue only when the full collect is finished.

Note that this change does not fix the bug where the board is not loaded
when it was closed while it was created at the same time.

But with this change, it's more easier to reproduce the bug because when
closing the board-search; it happens sometimes that a board-search is
not deleted properly (it stays in the state).

This board-search is not collected because he is its own owner. It
suggests that the remove must be sent explicitly for it and sometimes
it's not the case.