Try to fix definitivly the handling of subscriptions in the feeds
When a feed has more than one createdBy on a branch, the subscription
must continue to exists. The bug here was a special case where the
oldCreatedBy has an entry with a subscription for the feed, and the new
createdBy has still (at least) one subscription for the same branch.
In this case, the subscription will survive.