Prevent orphan entries in feeds when the parent exists in an other feed
There is a small check to be sure that this parent exists in an other
feed. We should check that the garbage collector works as expected. The
parent in the other feed should not be collected if at least one child
exists in an other feed (and I'm not sure that it's implemented).