Enable auto release with unsubscribeBranch and goblin-orc@
When a goblin-orc is dead, the branches attached to this
goblin-orc are auto-released. The reason is just because
the client node is no longer available, then it's impossible
to receive a warehouse.acknowledge in this case. The auto
release prevents leaks in the warehouse state in the case
of network interruptions.