Add new graft quest
It can be used in order to copy a branch from
one feed to an other feed. It follows the parents
and copy recursively all parents.

The children are not copied (of course) and the
copy is a merge. If the branch in the destination
already exists, it's just merged.