Rewrite the patch system in order to patch by branch instead of feed
The patches are generated against the branches when a branch changes for
a feed. It continues to use a throttle of 100 ms, and this throttle is
excited when data are upserted, feeds have changed or just deleted.

Because the patches are generated against the branches, it's no longer
necessary to loop into the whole feed subscriptions. A second point is
that the redux store subscription is no longer necessary.

The previous implementation was simpler but (it needs to be confirmed)
more CPU intensive.